Local functional test

+++ emonocot-portal/src/test/java/org/emonocot/test/TestDataManager.java
@@ -91,6 +91,8 @@ public class TestDataManager {
     private String username;
 
     private String password;
+    
+    private String useRemoteSolr;
 
     /**
      * @throws IOException
@@ -104,6 +106,7 @@ public class TestDataManager {
         properties.load(propertiesFile.getInputStream());
         username = properties.getProperty("functional.test.username", null);
         password = properties.getProperty("functional.test.password", null);
+        useRemoteSolr = properties.getProperty("use.remote.solr", "true");
     }
 
     @Autowired
@@ -971,19 +974,21 @@ public class TestDataManager {
 	}
 
 	public void cleanIndices() throws Exception {
-		ModifiableSolrParams params = new ModifiableSolrParams();
-    	params.add("q","*:*");
-    	params.add("df", "id");
-    	QueryResponse queryResponse = solrServer.query(params);
-    	SolrDocumentList solrDocumentList = queryResponse.getResults();
-    	List<String> documentsToDelete = new ArrayList<String>();
-    	for(int i = 0; i < solrDocumentList.size(); i++) {
-    		documentsToDelete.add(solrDocumentList.get(i).getFirstValue("id").toString());
-    	}
-    	if(!documentsToDelete.isEmpty()) {
-    	    solrServer.deleteById(documentsToDelete);
-    	    solrServer.commit(true,true);
-    	}
+		if(useRemoteSolr == "true") {
+		    ModifiableSolrParams params = new ModifiableSolrParams();
+    	    params.add("q","*:*");
+    	    params.add("df", "id");
+    	    QueryResponse queryResponse = solrServer.query(params);
+    	    SolrDocumentList solrDocumentList = queryResponse.getResults();
+    	    List<String> documentsToDelete = new ArrayList<String>();
+    	    for(int i = 0; i < solrDocumentList.size(); i++) {
+    		    documentsToDelete.add(solrDocumentList.get(i).getFirstValue("id").toString());
+    	    }
+    	    if(!documentsToDelete.isEmpty()) {
+    	        solrServer.deleteById(documentsToDelete);
+    	        solrServer.commit(true,true);
+    	    }
+		}
 	}
 
 	public void cleanDatabase() {
